HIV Classification
Making some progress, got the network training to converge by lowering
the training step.
Now I am having a problem with the mapping of the positive/negative
result.
I tries 0 and 1, but it seems the average for my network is around 0.5
for both positive and negative predictions - AKA is most likely not
learning anything.